@@ -13,28 +13,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^13.3.11",
-    "@angular/core": "^13.3.11",
-    "@angular/fire": "^5.2.3",
-    "@angular/forms": "^13.3.11",
-    "@angular/platform-browser": "^13.3.11",
-    "@angular/platform-browser-dynamic": "^13.3.11",
-    "@angular/router": "^13.3.11",
-    "@angular/service-worker": "^13.3.11",
-    "@awesome-cordova-plugins/core": "^5.45.0",
+    "@angular/common": "^18.2.14",
+    "@angular/core": "^18.2.14",
+    "@angular/fire": "^20.0.1",
+    "@angular/forms": "^18.2.14",
+    "@angular/platform-browser": "^18.2.14",
+    "@angular/platform-browser-dynamic": "^18.2.14",
+    "@angular/router": "^18.2.14",
+    "@angular/service-worker": "^18.2.14",
+    "@awesome-cordova-plugins/clipboard": "^8.1.0",
+    "@awesome-cordova-plugins/core": "^5.46.0",
+    "@awesome-cordova-plugins/file": "^8.1.0",
     "@awesome-cordova-plugins/file-opener": "^5.45.0",
-    "@ionic-native/clipboard": "^5.21.5",
-    "@ionic-native/core": "^5.0.0",
-    "@ionic-native/file": "^5.27.0",
-    "@ionic-native/file-transfer": "^5.28.0",
-    "@ionic-native/http": "^5.27.0",
-    "@ionic-native/httpd": "^5.27.0",
-    "@ionic-native/splash-screen": "^5.0.0",
-    "@ionic-native/status-bar": "^5.0.0",
-    "@ionic-native/zip": "^5.27.0",
-    "@ionic/angular": "^6.1.11",
+    "@awesome-cordova-plugins/file-transfer": "^8.1.0",
+    "@awesome-cordova-plugins/http": "^8.1.0",
+    "@awesome-cordova-plugins/splash-screen": "^8.1.0",
+    "@awesome-cordova-plugins/status-bar": "^8.1.0",
+    "@awesome-cordova-plugins/zip": "^8.1.0",
+    "@ionic/angular": "^7.8.6",
     "@ionic/cordova-builders": "^7.0.0",
-    "@ionic/storage": "^2.3.1",
+    "@ionic/storage": "^4.0.0",
+    "@ionic/storage-angular": "^4.0.0",
     "bootstrap": "^4.4.1",
     "buffer": "^5.2.1",
     "cordova-clipboard": "^1.3.0",
@@ -47,20 +46,23 @@
     "cordova-sqlite-storage": "^3.2.1",
     "cordova-support-android-plugin": "^1.0.1",
     "core-js": "^2.5.4",
+    "firebase": "^12.7.0",
     "font-awesome": "^4.7.0",
-    "rxjs": "^6.6.7",
+    "localforage-cordovasqlitedriver": "^1.8.0",
+    "rxjs": "^7.8.2",
+    "swiper": "^12.0.3",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.1303.8",
-    "@angular-devkit/build-angular": "~13.3.8",
-    "@angular-devkit/core": "~13.3.8",
-    "@angular-devkit/schematics": "~13.3.8",
-    "@angular/cli": "~13.3.8",
-    "@angular/compiler": "~13.3.11",
-    "@angular/compiler-cli": "~13.3.11",
-    "@angular/language-service": "~13.3.11",
+    "@angular-devkit/architect": "^0.1802.21",
+    "@angular-devkit/build-angular": "^18.2.21",
+    "@angular-devkit/core": "^18.2.21",
+    "@angular-devkit/schematics": "^18.2.21",
+    "@angular/cli": "^18.2.21",
+    "@angular/compiler": "^18.2.14",
+    "@angular/compiler-cli": "^18.2.14",
+    "@angular/language-service": "^18.2.14",
     "@ionic/angular-toolkit": "^7.0.0",
     "@ionic/lab": "2.0.6",
     "@types/jasmine": "~3.6.0",
@@ -83,7 +85,7 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.4.2"
+    "typescript": "~5.4.5"
   },
   "description": "An Ionic project",
   "cordova": {
@@ -108,5 +110,8 @@
     "platforms": [
       "android"
     ]
+  },
+  "optionalDependencies": {
+    "@rollup/rollup-linux-arm64-gnu": "^4.54.0"
   }
 }
