@@ -1,8 +1,12 @@
 import { Component } from '@angular/core';
+import { register } from 'swiper/element/bundle';
+
+register();
 
 import { Platform } from '@ionic/angular';
-import { SplashScreen } from '@ionic-native/splash-screen/ngx';
-import { StatusBar } from '@ionic-native/status-bar/ngx';
+import { SplashScreen } from '@awesome-cordova-plugins/splash-screen/ngx';
+import { StatusBar } from '@awesome-cordova-plugins/status-bar/ngx';
+import { Storage as IonicStorage } from '@ionic/storage-angular';
 
 @Component({
   selector: 'app-root',
@@ -12,12 +16,14 @@ export class AppComponent {
   constructor(
     private platform: Platform,
     private splashScreen: SplashScreen,
-    private statusBar: StatusBar
+    private statusBar: StatusBar,
+    private storage: IonicStorage
   ) {
     this.initializeApp();
   }
 
-  initializeApp() {
+  async initializeApp() {
+    await this.storage.create();
     this.platform.ready().then(() => {
       this.statusBar.styleDefault();
       this.splashScreen.hide();
@@ -25,10 +31,10 @@ export class AppComponent {
     });
   }
 
-  changeDarkMode(){
+  changeDarkMode() {
     const prefersDark = window.matchMedia('(prefers-color-scheme: dark)');
-    if (prefersDark.matches){
-    document.body.classList.toggle('dark');
-      }
+    if (prefersDark.matches) {
+      document.body.classList.toggle('dark');
     }
+  }
 }
