@@ -1,4 +1,4 @@
-import { NgModule } from '@angular/core';
+import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { Routes, RouterModule } from '@angular/router';
@@ -21,6 +21,7 @@ const routes: Routes = [
     IonicModule,
     RouterModule.forChild(routes)
   ],
-  declarations: [PlanDetallePage]
+  declarations: [PlanDetallePage],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
-export class PlanDetallePageModule {}
+export class PlanDetallePageModule { }
